code/sml-rw/requirements.txt
colorama==0.4.6
loguru==0.7.2
pyserial==3.5
smllib==1.3
win32-setctime==1.1.0
pyftdi==0.55.0
fire==0.5.0
pyusb==1.2.1